requirements.txt
httpx>=0.18.2
requests>=2.25.1
opencv-python>=4.5.4.58
numpy>=1.22.3
unidecode>=1.3.4
onnx>=1.11.0
onnxruntime>=1.9.0
pyyaml>=5.4.1
pycryptodome==3.10.1